+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/CheckboxPicker/CheckboxPickerConfiguration.swift
@@ -136,16 +136,18 @@ struct CheckboxPickerConfiguration: View {
 
     var body: some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedMedium) {
-            OUDSSwitchItem("app_common_enabled_label", isOn: $model.isEnabled)
-                .disabled(model.isError || model.isReadOnly)
+            VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
+                OUDSSwitchItem("app_common_enabled_label", isOn: $model.isEnabled)
+                    .disabled(model.isError || model.isReadOnly)
 
-            OUDSSwitchItem("app_components_common_error_label", isOn: $model.isError)
-                .disabled(!model.isEnabled || model.isReadOnly)
+                OUDSSwitchItem("app_components_common_error_label", isOn: $model.isError)
+                    .disabled(!model.isEnabled || model.isReadOnly)
 
-            OUDSSwitchItem("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
-                .disabled(!model.isEnabled || model.isError)
+                OUDSSwitchItem("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
+                    .disabled(!model.isEnabled || model.isError)
 
-            OUDSSwitchItem("app_components_controlItem_divider_label", isOn: $model.hasDivider)
+                OUDSSwitchItem("app_components_controlItem_divider_label", isOn: $model.hasDivider)
+            }
 
             DesignToolboxChoicePicker(title: "app_components_common_orientation_label",
                                       selection: $model.pickerPlacement,