+++ DesignToolbox/DesignToolbox/Pages/Components/ControlItem/ControlItemConfigurationModel.swift
@@ -14,19 +14,21 @@
 import OUDSComponents
 import SwiftUI
 
-// MARK: - Checkbox Item Configuration Model
+// MARK: - Control Item Configuration Model
 
-/// The model shared between `CheckboxItemConfiguration` view and `CheckboxItemPage` view.
-/// Related to `OUDSCheckboxItem` (i.e. with 2 available values).
-final class CheckboxItemConfigurationModel: ComponentConfiguration {
+/// The model shared between `ControlItemConfigurationModel` view and `ControlItemConfiguration` view.
+class ControlItemConfigurationModel: ComponentConfiguration {
+
+    typealias OutlinedConfiguration = (value: Bool, outlinedConfigurationLabel: String)
+    typealias AdditionalLabelConfiguration = String
 
     // MARK: - Properties
 
-    @Published var enabled: Bool {
-        didSet { updateCode() }
-    }
+    var componentInitCode: String = ""
+    var additionalLabelConfiguration: AdditionalLabelConfiguration?
+    var outlinedConfiguration: OutlinedConfiguration?
 
-    @Published var indicatorState: Bool {
+    @Published var enabled: Bool {
         didSet { updateCode() }
     }
 
@@ -62,19 +64,34 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
+    @Published var outlined: Bool {
+        didSet { updateCode() }
+    }
+
+    @Published var additionalLabelText: String {
+        didSet { updateCode() }
+    }
+
     // MARK: - Initializer
 
-    override init() {
-        indicatorState = true
+    init(componentInitCode: String,
+         outlinedConfiguration: OutlinedConfiguration? = nil,
+         additionalLabelConfiguration: AdditionalLabelConfiguration? = nil)
+    {
+        self.componentInitCode = componentInitCode
         isError = false
         isReadOnly = false
         enabled = true
         icon = true
         flipIcon = false
         isReversed = false
-        divider = false
+        divider = true
         labelText = String(localized: "app_components_common_label_label")
         helperText = String(localized: "app_components_controlItem_helperText_label")
+        self.outlinedConfiguration = outlinedConfiguration
+        self.additionalLabelConfiguration = additionalLabelConfiguration
+        outlined = outlinedConfiguration?.value ?? false
+        additionalLabelText = additionalLabelConfiguration ?? ""
     }
 
     deinit {}
@@ -85,15 +102,15 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
     override func updateCode() {
         code =
             """
-            OUDSCheckboxItem(isOn: $isOn, label: \"\(labelText)\"\(helperTextPattern)\(iconPattern)\(flipIconPattern)\(isReversedPattern)\(isErrorPattern)\(isReadOnlyPattern)\(dividerPattern))
-            \(disableCode)
+            \(componentInitCode), label: "\(labelText)"\(additionalLabelTextPattern)\(helperTextPattern)\(iconPattern)\(flipIconPattern)\(outlinedPattern)\(isReversedPattern)\(isErrorPattern)\(isReadOnlyPattern)\(dividerPattern))
+            \(disableCodePattern)
             """
     }
 
     // swiftlint:enable line_length
 
-    private var disableCode: String {
-        ".disabled(\(enabled ? "false" : "true"))"
+    private var disableCodePattern: String {
+        !enabled ? ".disabled(true)" : ""
     }
 
     private var helperTextPattern: String {
@@ -109,7 +126,7 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
     }
 
     private var isReversedPattern: String {
-        ", isReversed: \(isReversed)"
+        isReversed ? ", isReversed: true" : ""
     }
 
     private var isErrorPattern: String {
@@ -123,42 +140,28 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
     private var dividerPattern: String {
         divider ? ", divider: true" : ""
     }
+
+    private var additionalLabelTextPattern: String {
+        !additionalLabelText.isEmpty ? ", additionalLabel: \"\(additionalLabelText)\"" : ""
+    }
+
+    private var outlinedPattern: String {
+        outlined == true ? ", isOutlined: true" : ""
+    }
 }
 
-// MARK: - Checkbox Item Configuration View
+// MARK: - ControlItem Configuration View (Bool selection)
 
-struct CheckboxItemConfiguration: View {
+final class BooleanControlItemConfigurationModel: ControlItemConfigurationModel {
+    @Published var isOn: Bool = true
 
-    @ObservedObject var model: CheckboxItemConfigurationModel
+    deinit {}
+}
 
-    @Environment(\.theme) private var theme
+// MARK: - ControlItem Configuration View (Indeterminate checkbox selection)
 
-    var body: some View {
-        VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
-            OUDSSwitchItem("app_components_checkbox_selection_label", isOn: $model.indicatorState)
+final class IndicatorControlItemConfigurationModel: ControlItemConfigurationModel {
+    @Published var selection: OUDSCheckboxIndicatorState = .indeterminate
 
-            OUDSSwitchItem("app_components_controlItem_icon_label", isOn: $model.icon)
-
-            OUDSSwitchItem("app_components_controlItem_flipIcon_label", isOn: $model.flipIcon)
-                .disabled(!model.icon)
-
-            OUDSSwitchItem("app_components_controlItem_divider_label", isOn: $model.divider)
-
-            OUDSSwitchItem("app_components_controlItem_reversed_label", isOn: $model.isReversed)
-
-            OUDSSwitchItem("app_common_enabled_label", isOn: $model.enabled)
-                .disabled(model.isError || model.isReadOnly)
-
-            OUDSSwitchItem("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
-                .disabled(!model.enabled || model.isError)
-
-            OUDSSwitchItem("app_components_common_error_label", isOn: $model.isError)
-                .disabled(!model.enabled || model.isReadOnly)
-        }
-
-        DesignToolboxEditContentDisclosure {
-            DesignToolboxTextField(text: $model.labelText)
-            DesignToolboxTextField(text: $model.helperText)
-        }
-    }
+    deinit {}
 }