+++ DesignToolbox/DesignToolbox/Pages/Components/Switch/Switchitem/SwitchItemPage.swift
@@ -19,47 +19,11 @@ import SwiftUI
 
 struct SwitchItemPage: View {
 
-    @StateObject private var configuration: SwitchItemConfigurationModel
-
-    init() {
-        _configuration = StateObject(wrappedValue: SwitchItemConfigurationModel())
-    }
+    let model = BooleanControlItemConfigurationModel(componentInitCode: "OUDSSwitchItem(isOn: $isOn")
 
     var body: some View {
-        ComponentConfigurationView(
-            configuration: configuration,
-            componentView: componentView,
-            configurationView: configurationView)
-    }
-
-    @ViewBuilder
-    private func componentView(with configuration: ComponentConfiguration) -> some View {
-        if let model = configuration as? SwitchItemConfigurationModel {
-            SwitchItemIllustration(model: model)
-        }
-    }
-
-    @ViewBuilder
-    private func configurationView(with configuration: ComponentConfiguration) -> some View {
-        if let model = configuration as? SwitchItemConfigurationModel {
-            SwitchItemConfiguration(model: model)
-        }
-    }
-}
-
-// MARK: - Switch Item Illustration
-
-private struct SwitchItemIllustration: View {
-
-    let model: SwitchItemConfigurationModel
-    @Environment(\.colorScheme) private var colorScheme
-
-    var body: some View {
-        VStack(alignment: .center) {
-            // TODO: Build a modifier to inverse colorscheme or force to a colorscheme
+        ControlItemElementPage(model: model) {
             SwitchItemDemo(model: model)
-            SwitchItemDemo(model: model)
-                .colorScheme(colorScheme == .dark ? .light : .dark)
         }
     }
 }
@@ -68,12 +32,12 @@ private struct SwitchItemIllustration: View {
 
 private struct SwitchItemDemo: View {
 
-    @ObservedObject var model: SwitchItemConfigurationModel
+    @ObservedObject var model: BooleanControlItemConfigurationModel
     @Environment(\.theme) private var theme
 
     var body: some View {
         OUDSSwitchItem(model.labelText,
-                       isOn: $model.selection,
+                       isOn: $model.isOn,
                        helper: model.helperText,
                        icon: icon,
                        flipIcon: model.flipIcon,